Major changes to json files reading, developing launcher

+++ app/src/main/java/com/ne/revival_games/entity/WorldObjects/Entity/Aim/AimableEntity.java
@@ -50,5 +50,5 @@ public abstract class AimableEntity extends Entity implements Aimable {
         this.logic = logic;
     }
 
-    public abstract int getTurnSpeed();
+    public abstract double getTurnSpeed();
 }
